pyproject.toml
[project]
name = "pygments-keylex"
version = "0.1.0"
dependencies = [
    "Pygments",
    'antlr4-python3-runtime',
]

[build-system]
requires = ["setuptools >= 51.0.0"]
build-backend = "setuptools.build_meta"

[project.entry-points."pygments.lexers"]
key = "keylex:KeYLexer"
jml = "keylex:JMLLexer"

